fix: 학습 모델의 LateDropout 인풋 불일치 (#73)
* fix: 학습 모델의 ECA.call 인풋 일치

* fix: 학습 모델의 LateDropout 인풋 불일치

+++ src/backbone.py
@@ -100,9 +100,8 @@ class LateDropout(keras.layers.Layer):
         # dtype = inputs.dtype
         x = tf.cond(
             self._train_counter < self.start_step,
-            lambda:inputs,
-            # lambda: ops.cast(self.dropout(inputs, training=training), dtype)        # inputs의 dtype으로 casting - fp16 가속 설정
-            lambda: self.dropout(inputs, training=training)
+            lambda: inputs,
+            lambda: tf.cast(self.dropout(inputs, training=training), inputs.dtype)
         )
         if training:
             self._train_counter.assign_add(1)
